Add InferenceEngine object
