+++ src/prophetverse/engine/__init__.py
@@ -1,0 +1,12 @@
+"""Module for the inference engines."""
+
+from .base import BaseInferenceEngine
+from .map import MAPInferenceEngine, MAPInferenceEngineError
+from .mcmc import MCMCInferenceEngine
+
+__all__ = [
+    "BaseInferenceEngine",
+    "MAPInferenceEngine",
+    "MAPInferenceEngineError",
+    "MCMCInferenceEngine",
+]
